split auth form and auth page into their own components
